[AI] test(plugins): split admission-failure regression to its own file
CI check-lint flagged install-persistence.test.ts at 1015 effective lines
(eslint(max-lines), max 1000) after the admission-failure regression was
added in 80f898b748c. Move that test to a new install-persistence-admission.test.ts
so the original file drops back under the limit. Test content unchanged.
