+++ src/plugins/install-persistence.test.ts
@@ -941,51 +941,6 @@ describe("persistPluginInstall", () => {
     expect(writeConfigFile).not.toHaveBeenCalled();
   });
 
-  it("rejects a plugin whose manifest admission failed instead of persisting it as ready", async () => {
-    const { persistPluginInstall } = await import("./install-persistence.js");
-    const baseConfig = {
-      plugins: {
-        allow: ["memory-core"],
-        entries: {},
-      },
-    } as OpenClawConfig;
-    // A structurally invalid configSchema is rejected at manifest admission: the plugin
-    // is absent from registry.plugins and appears only as an error diagnostic. The install
-    // path must not map that absence to "ready" and write a successful install record.
-    loadPluginManifestRegistry.mockReturnValue({
-      plugins: [],
-      diagnostics: [
-        {
-          level: "error",
-          pluginId: "broken-schema-plugin",
-          message:
-            "plugin manifest configSchema is invalid: <schema>.properties.mode.$ref: unresolved ref",
-          source: "/tmp/broken-schema-plugin/openclaw.plugin.json",
-        },
-      ],
-    });
-
-    await expect(
-      persistPluginInstall({
-        snapshot: {
-          config: baseConfig,
-          baseHash: "config-1",
-          writeOptions: installWriteOptions,
-        },
-        pluginId: "broken-schema-plugin",
-        install: {
-          source: "npm",
-          spec: "broken-schema-plugin@1.0.0",
-          installPath: "/tmp/broken-schema-plugin",
-        },
-      }),
-    ).rejects.toThrow("has invalid configured settings");
-
-    expect(enablePluginInConfig).not.toHaveBeenCalled();
-    expect(writePersistedInstalledPluginIndexInstallRecordsWithLease).not.toHaveBeenCalled();
-    expect(writeConfigFile).not.toHaveBeenCalled();
-  });
-
   it("can persist an install record without enabling a plugin that needs config first", async () => {
     const { persistPluginInstall } = await import("./install-persistence.js");
     const baseConfig = {
